[CIR][CodeGen] introduce CIRBaseBuilder
